Emit correct save/restore sequence for callee saved vectors
Our experimental epi_vectorcall calling convention reserves a number of
vector registers as callee saves (the idea is that a function that calls
an epi_vectorcall one does not have to save and restore all its live
vectors around the call).

However, this comes as a downside that we may have callee saved of
vector registers, which is not well supported in the RISC-V backend
because vector registers are always caller saved. What happens is that
we save them, but too early, and then we don't restore them. All this is
related to the StackID. Initially the StackID is set to default but we
change it during frame lowering to the ScalableStack. This means that
one of the scans through the FrameInfos does include the vector
registers while the second does not. This is only a problem if we have
callee saved registers of vector type, which by default we do not.

This change fixes this problem by adding an extra hook in
TargetFrameLowering so the target has a chance to set the StackID of
callee saves before frame lowering starts. It does nothing by default
but the RISC-V backend will use this to set the StackID of vector callee
saved to StackID::ScalableVector.

This way we can keep ignoring those non-default StackID callee saved
during {save,restore} as it happens already, and then handle them
manually {after,before} the storage for them (which was already
correctly accounted) has been {allocated,freed}.