Add cases for Predicated VPRecipes in getter
These two getters were incomplete, and mayHaveSideEffects was returning
(conservatively) true for VPPredicatedFirstOrderRecurrence which caused
the creation of a VPlan that when destroyed failed because of a VPValue
was still had VPUser (weirdly an IR value that came as a live-in so this
feels like a true issue).

Theoretically the VPlan may still be wrong and we should fix this
problem but for now let's add the missing machinery and get the
testsuite compiling again.